Don't overwrite .emacs

+++ install.bash
@@ -11,5 +11,11 @@ fi
 
 mkdir ~/.emacs.d                     2>> /dev/null
 ln -s ~/dev/dotemacs ~/elisp         2>> /dev/null
-ln -s ~/dev/dotemacs/.emacs ~/.emacs 2>> /dev/null
+
+if [ -e ~/.emacs ]; then
+    echo "~/.emacs exists, run 'ln -s ~/dev/dotemacs/.emacs ~/.emacs' to overwrite it"
+else
+    ln -s ~/dev/dotemacs/.emacs ~/.emacs 2>> /dev/null
+fi
+
 ln -s ~/elisp/elpa ~/.emacs.d/elpa   2>> /dev/null
